Split models in two separated files
