+++ README.md
@@ -13,8 +13,7 @@ En cas de problème avec l'utilisation de la base ou pour toute demande d’info
 pantheonisees
     ├── app
     │   ├── modeles
-    │   │       ├── images.py
-    │   │       ├── pantheonisees.py
+    │   │       ├── data.py
     │   │       └── user.py
     │   ├── routes
     │   │       ├── connexion.py
