+++ pantheonisees/app/app.py
@@ -61,6 +61,5 @@ db = SQLAlchemy(app)
 
 # Imports locaux
 from .modeles.user import *
-from .modeles.pantheonisees import *
-from .modeles.images import *
+from .modeles.data import *
 from .routes import general, search, errors, connexion, crud_person
