+++ pantheonisees/app/routes/search.py
@@ -2,7 +2,7 @@ from flask import render_template, request
 from sqlalchemy import and_
 
 from ..app import *
-from ..modeles.pantheonisees import *
+from ..modeles.data import *
 
 
 @app.route("/recherche")
